src/requirements.txt
fastapi
uvicorn
pydantic
arxiv
langchain-community

fastapi
uvicorn
pydantic
numexpr


fastapi
uvicorn
pydantic
langchain-core
langchain-community
rizaio

fastapi
uvicorn
pandas
python-multipart

fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-core
langchain-ollama
ipython


fastapi
uvicorn
pydantic
langchain-core
python-dotenv
langchain-ollama
langgraph
ipython
langchain

